refactor: extract tm_dynamic logic to genhcl
